go/types, types2: AssertableTo cannot be used for generalized interfaces
Document that this is not supported (at least for 1.18) and that
the function panics if called with generalized interfaces.

While at it, also removed the internal flag forceStrict and its
one use in Checker.assertableTo; forceStrict was never enabled
and if it would have been enabled, the behavior would not have
been correct.

Change-Id: Ie4dc9345c88d04c9640f881132154a002db22643