runtime: refactor the scavenger and make it testable
DO NOT SUBMIT

Needs tests.

This change refactors the scavenger into a type whose methods represent
the actual function and scheduling of the scavenger. It also stubs out
access to global state in order to make it testable.

While we're here, I realized that there's no real reason wakeScavenger
(now scavenger.wake) needs to be the routine that stops the timer. It
just needs to wake up the scavenger itself. It risks double wake-ups,
but we already had those before and we're resilient to them. Given
that, we can remove the whole sysmon mechanism.

Change-Id: Ib31b86379fd9df84f25ae282734437afc540da5c