io/fs: add ReadLinkFS interface
Add implementations for *io/fs.subFS and os.DirFS,
and check behavior in testing/fstest.TestFS.

I am deliberately not changing archive/zip in this CL,
since it currently does not resolve symlinks
as part of its filesystem implementation.
I am unsure of the compatibility restrictions on doing so,
so figured it would be better to address independently.

Fixes #49580

Change-Id: I58ec6915e8cc97341cdbfd9c24c67d1b60139447