runtime: add byte count parser for GOMEMLIMIT
This change adds a parser for the GOMEMLIMIT environment variable's
input. This environment variable accepts a number followed by an
optional prefix expressing the unit. Acceptable units include
B, KB, KiB, MB, MiB, GB, GiB, TB, TiB, where *iB is a power-of-two byte
unit and *[^i]B is a power-of-ten unit.

For #48409.

Change-Id: I6a3b4c02b175bfcf9c4debee6118cf5dda93bb6f