go/types, types2: implement (static) IsComparable predicate
A type implements a comparable interface only if the type
is statically known to be comparable. Specifically, a type
cannot cannot contain (component) interfaces that are not
statically known to be comparable.

This CL implements a new predicate _IsComparable which is
like Comparable, but it only reports true for an interface
if the interface is statically known to be comparable.

To implement _IsComparable, rename the existing (internal)
comparable predicate to isComparable (for symmetry) and
add an additional argument to indicate that an interface's
comparability may only be known dynamically.

Fixes #51257.

Change-Id: If22bc047ee59337deb2e7844b8f488d67e5c5530