cmd/compile: prototype for dealing with constructed types with shapes
This is a different way to deal with the problem of constructed types
with shapes in them. In this approach, we roughly remove all named types
from the entire shape of a shape type. We do that for all shape types
created from normal/concrete types, and also when creating shape types
from constructed types in instantiations that have shapes in them. By
removing all named types (except for recursion), we get a more canonical
type and automatically deal with shape types inside a constructed.

The problem is that we need to adjust types.Identity to so that the
shape types still compare identical with the original non-shape type.
Also, it still doesn't work in issue23536.go, where inlining is hiding
the shape type, so even extra flags/checks in types.Identity don't help
(so *MyByte and *byte don't match).

Change-Id: I2a6938d694bb16dc50ac54235bf1d94de3da2ade