time/tzdata: use go:embed for zoneinfo.zip
The package included generate_zipdata.go,
which took care of manually embedding the zip file as a string constant.
That's exactly what go:embed exists for nowadays,
so use that directly and remove all that code.

Note that embedding does not allow referencing parent directories;
one can't use //go:embed ../../../lib/time/zoneinfo.zip.
For that reason, move the zip file inside src/time/tzdata.

And, arguably, that's a good thing.
Before this change, src/time/tzdata already contained the zip once,
via the code-generated tzdata.go. It would be perhaps a bit extra
for the same directory to contain the same file in two forms.
The original zip file was also unnecessary to build std,
so one could "rm -rf lib" and Go could still work, saving space.

Now that we no longer duplicate the zip file in a second form,
and now that we embed the file directly,
move the zip file from lib/time to src/time/tzdata.
For consistency and ease, we move update.bash and README too.

Finally, a few bits of code referenced the old zip path,
to be used as a fallback when loading up timezone information.
Update those to reference the new path under GOROOT.

Change-Id: I3302cd4d0d10be0a2e66f7672e96b77cb6e37d28