runtime: don't unmap arena chunk address regions, make them fault instead
Right now arena chunks are completely unmapped in order to force a fault
on any dereferences of pointers to those regions. However, a hard
unmapping means that at any point the process could map something else
into that region, breaking the safety guarantees.

Instead, make the region as the equivalent of PROT_NONE (i.e. set it up
to fault) and mark the region as unused to release any memory backing
it.

Change-Id: Id1475548bd3c10dbf5672e0d85333193ca9c18c3