runtime: recycle arena address space when no more pointers are found
This change modifies the arena implementation to retain the mspan and
heapArena structures in place indefinitely, enabling us to track when
all pointers into the arena finally go away. This lets us safely reuse
the arena.

The previous implementation would hide the address space from the GC,
and hand over the data structures. Part of the purpose of this is so
that live pointers to those regions don't cause faults after an arena is
freed, provided that said pointer doesn't get dereferenced. The address
space thus must be hidden from the GC, so that the GC doesn't fault.

As said before, we keep the data structures in place, but to prevent the
fault from the GC, we mark the area as noscan. The GC will still mark
the arena, but it will no longer try to scan it while it remains freed.
Once we fail to see it marked, we can safely reuse the span.

Change-Id: I5a4c94330839e49bd8082a8b06569e2f6634ea90