cmd/compile: modify switches of strings to use jump table for lengths
Reorganize the way we rewrite expression switches on strings, so that
jump tables are naturally used for the outer switch on the string length.

The changes to the prove pass in this CL are required so as to not repeat
the test for string length in each case.

name                         old time/op  new time/op  delta
SwitchStringPredictable-8    1.99ns ± 6%  1.98ns ± 4%     ~     (p=0.646 n=10+9)
SwitchStringUnpredictable-8  11.3ns ± 2%   8.3ns ±13%  -26.19%  (p=0.000 n=9+10)

Update #5496
Update #34381

Change-Id: Ie6846b1dd27f3e472f7c30dfcc598c68d440b997