cmd/compile: add coverage fixup mode
DO NOT SUBMIT

Adds a -coveragecfg=<configfile> command line option to the compiler
to help support a cooperative "tool and compiler" mode for coverage
instrumentation. In this mode the cmd/cover tool generates most of the
counter instrumentation via source-to-source rewriting, but the
compiler fixes up the result if passed the "-coveragecfg" option. The
fixups include:

  - reclassifying counter variables (special storage class)
  - marking meta-data variables are read-only
  - adding in an init call to do registation

Notes:
- probably need to also make changes to ensure that we don't
  generate DWARF entries for counter, meta-data, and pkgid
  variables

Change-Id: Iead72b85209725ee044542374465f118a3ee72e0